reuse frontmatter extraction
