+++ lib/generate-blog-components/lib/collect-posts.js
@@ -4,8 +4,8 @@
 const path = require('path');
 
 const glob = require('glob');
-const fs = require('fs-extra');
-const frontMatter = require('front-matter');
+
+const separateMeta = require('../../markdown-content/separate-meta');
 
 module.exports = function(folder) {
   return glob
@@ -40,18 +40,4 @@ function buildComponentName(file) {
     .replace(/[^a-zA-Z0-9]/, '_');
 
   return `BlogPost${cleanedFileName}`;
-}
-
-function separateMeta(file) {
-  let content = fs.readFileSync(file).toString();
-  let data = frontMatter(content);
-  let date = new Date(path.basename(file).substring(0, 10));
-  let [excerpt, body] = data.body.split('<!--break-->');
-
-  let meta = {
-    ...data.attributes,
-    date,
-  };
-
-  return [meta, excerpt, body];
 }
