+++ lib/generate-calendar/lib/collect-events.js
@@ -4,8 +4,8 @@
 const path = require('path');
 
 const glob = require('glob');
-const fs = require('fs-extra');
-const frontMatter = require('front-matter');
+
+const separateMeta = require('../../markdown-content/separate-meta');
 
 module.exports = function(folder) {
   return glob
@@ -22,17 +22,3 @@ module.exports = function(folder) {
       };
     });
 };
-
-function separateMeta(file) {
-  let content = fs.readFileSync(file).toString();
-  let data = frontMatter(content);
-  let date = new Date(path.basename(file).substring(0, 10));
-  let [excerpt, body] = data.body.split('<!--break-->');
-
-  let meta = {
-    ...data.attributes,
-    date,
-  };
-
-  return [meta, excerpt, body];
-}
